fix special item data imports
