+++ src/components/modals/AddStoneModal.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
 import { useTranslation } from "react-i18next";
 import { STONES } from "@/src/data/stones.ts";
-import { MEGA_STONES } from "@/src/services/init";
+import { MEGA_STONES } from "@/src/data/megastones.ts";
 import { FiX, FiInfo } from "react-icons/fi";
 import ToggleSwitch from "@/src/components/toggles/ToggleSwitch";
 import LocationSuggestionInput from "../inputs/LocationSuggestionInput.tsx";
